💄 옵션 추가할 시 유동적으로 박스 크기 늘어나도록스타일링 구현 완료
